FIX handling of parameter names ending '_'
This issue results from #107: the explicit bold styling was removed
because Sphinx definition lists bolded the term by default. However,
without ** surrounding param name, a name like word_ was being treated
as a link. I also attempted to replace the _ with \_ but found that
the backslashes showed in the HTML. Thus I have reverted to wrapping
param names in **.

Conflicts:
	numpydoc/docscrape_sphinx.py
	numpydoc/tests/test_docscrape.py